rename for consistency: type_script
